Only create Slack channel on demand
We no longer create the Slack Channel on each PR, we only create it when
you either click the link in the 'managed comment' or if you say the
magic take it to slack command.

We also introduce the concept of a single 'managed comment' so that we
can have the potential to update this comment, if things change.
Before we only made this comment when we were creating the Slack
channel, but now we want this comment to exist first.
Before we were 'cheating' and using the Slack Channel existing as the
state in determining whether or not to leave this comment.
Now we are using the comment itself for this state, which is good for
being stateless! We add a special HTML comment to the top with a key we
can search for to find this comment later