Add extra new lines before PR footer
When Live Github adds the 'footer' to each PR it didn't add any new
lines between the 'real' content and our future comments. This was a
common annoyance for me when I went to edit my Github PR descriptions,
so decided to fix it today!

This simply adds two new lines between the original body and our footer
when we insert it